Let user specify search path for .reviewboardrc
If a user sets $RBTOOLS_CONFIG_PATH, it should contain a list of paths
separated by the characters conventionally used for this on the
operating system (":" on Linux, Mac OS X, ";" on Windows). These paths
are searched first for .reviewboardrc and can be used to override
repository-wide settings.

Reviewed at https://reviews.reviewboard.org/r/7679/